feat: add Golang 1.23.1 with asdf-vm
go1.23.0 (released 2024-08-13) is a major release of Go.

https://go.dev/doc/devel/release#go1.23.0

go1.23.1 (released 2024-09-05) includes security fixes to the encoding/gob,
go/build/constraint, and go/parser packages, as well as bug fixes to the
compiler, the go command, the runtime, and the database/sql, go/types, os,
runtime/trace, and unique packages. See the Go 1.23.1 milestone on our
issue tracker for details.

https://github.com/golang/go/issues?q=milestone%3AGo1.23.1+label%3ACherryPickApproved

Signed-off-by: Stephan Linz <linz@li-pro.net>