feat: add Python 2.7.18 with asdf-vm
+++ OBSOLETE and DEPRECATED +++

For historical reasons only!

+++ OBSOLETE and DEPRECATED +++

Python 2.7.18 is the last release of Python 2.

The End Of Life date (EOL, sunset date) for Python 2.7 has been
moved five years into the future, to 2020. Python 2.7 is the end
of the Python 2.x series, and is succeeded by Python 3.

Being the last of the 2.x series, 2.7 received bugfix support
until 2020. Support officially stopped January 1 2020, and
2.7.18 code freeze occurred on January 1 2020, but the final
release occurred after that date.

https://peps.python.org/pep-0373
https://www.python.org/downloads/release/python-2718

Signed-off-by: Stephan Linz <linz@li-pro.net>