fix(ci)!: beeing under limit of 6 hours build time
+++ CORRECT THIS AS SOON AS POSSIBLE! +++

In general, the maximum runtime of jobs in the GitHub CI workflow
is limited to 6 hours and can only be increased on a customer-specific
basis in the case of self-hosted runners. As long as no matrix strategy
is used and therefore each platform is built separately in individual
jobs, the number of platforms must be "artificially" limited.

+++ CORRECT THIS AS SOON AS POSSIBLE! +++

Signed-off-by: Stephan Linz <linz@li-pro.net>